requirements.txt
Django==6.0.3
django-allauth==0.57.2
channels==4.0.0
daphne==4.0.0
channels-redis>=4.0.0
Pillow>=10.4.0
python-decouple==3.8
requests==2.31.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
whitenoise==6.6.0
python-dotenv==1.0.0
certifi>=2024.2.2
openai>=1.0.0
razorpay>=1.0.0
numpy
scikit-learn